create a QgsDatumTransformTableWidget to avoid code duplication
this contains the table view and the 3 buttons (add/remove/edit)
