+++ src/app/CMakeLists.txt
@@ -22,7 +22,7 @@ SET(QGIS_APP_SRCS
   qgsclipboard.cpp
   qgscustomization.cpp
   qgscustomprojectiondialog.cpp
-  qgsdatumtransformtablemodel.cpp
+  qgsdatumtransformtablewidget.cpp
   qgsdecorationitem.cpp
   qgsdecorationcopyright.cpp
   qgsdecorationcopyrightdialog.cpp
@@ -247,7 +247,7 @@ SET (QGIS_APP_MOC_HDRS
   qgsclipboard.h
   qgscustomization.h
   qgscustomprojectiondialog.h
-  qgsdatumtransformtablemodel.h
+  qgsdatumtransformtablewidget.h
   qgsdecorationitem.h
   qgsdecorationcopyright.h
   qgsdecorationcopyrightdialog.h
