+++ src/app/qgsprojectproperties.cpp
@@ -24,7 +24,7 @@
 #include "qgisapp.h"
 #include "qgscomposer.h"
 #include "qgscoordinatetransform.h"
-#include "qgsdatumtransformdialog.h"
+#include "qgsdatumtransformtablewidget.h"
 #include "qgslayoutmanager.h"
 #include "qgslogger.h"
 #include "qgsmapcanvas.h"
@@ -157,15 +157,7 @@ QgsProjectProperties::QgsProjectProperties( QgsMapCanvas *mapCanvas, QWidget *pa
 
   // Datum transforms
   QgsCoordinateTransformContext context = QgsProject::instance()->transformContext();
-  mDatumTransformTableModel->setTransformContext( context );
-  mDatumTransformTableView->setModel( mDatumTransformTableModel );
-  mDatumTransformTableView->resizeColumnToContents( 0 );
-  mDatumTransformTableView->horizontalHeader()->show();
-  mDatumTransformTableView->setSelectionMode( QAbstractItemView::SingleSelection );
-  mDatumTransformTableView->setSelectionBehavior( QAbstractItemView::SelectRows );
-  connect( mDatumTransformAddButton, &QToolButton::clicked, this, &QgsProjectProperties::addDatumTransform );
-  connect( mDatumTransformRemoveButton, &QToolButton::clicked, this, &QgsProjectProperties::removeDatumTransform );
-  connect( mDatumTransformEditButton, &QToolButton::clicked, this, &QgsProjectProperties::editDatumTransform );
+  mDatumTransformTableWidget->setTransformContext( context );
 
   bool show = settings.value( QStringLiteral( "/Projections/showDatumTransformDialog" ), false ).toBool();
   mShowDatumTransformDialogCheckBox->setChecked( show );
@@ -822,7 +814,7 @@ void QgsProjectProperties::apply()
     projectionSelector->pushProjectionToFront();
   }
 
-  QgsCoordinateTransformContext transformContext = mDatumTransformTableModel->transformContext();
+  QgsCoordinateTransformContext transformContext = mDatumTransformTableWidget->transformContext();
   QgsProject::instance()->setTransformContext( transformContext );
 
   // Set the project title
@@ -1232,71 +1224,6 @@ void QgsProjectProperties::apply()
 void QgsProjectProperties::showProjectionsTab()
 {
   mOptionsListWidget->setCurrentRow( 1 );
-}
-
-void QgsProjectProperties::addDatumTransform()
-{
-  QgsDatumTransformDialog *dlg = new QgsDatumTransformDialog();
-  if ( dlg->exec() )
-  {
-    QPair< QPair<QgsCoordinateReferenceSystem, int>, QPair<QgsCoordinateReferenceSystem, int > > dt = dlg->selectedDatumTransforms();
-    QgsCoordinateTransformContext context = mDatumTransformTableModel->transformContext();
-    context.addSourceDestinationDatumTransform( dt.first.first, dt.second.first, dt.first.second, dt.second.second );
-    mDatumTransformTableModel->setTransformContext( context );
-  }
-}
-
-void QgsProjectProperties::removeDatumTransform()
-{
-  QModelIndexList selectedIndexes = mDatumTransformTableView->selectionModel()->selectedIndexes();
-  if ( selectedIndexes.count() > 0 )
-  {
-    mDatumTransformTableModel->removeTransform( selectedIndexes );
-  }
-}
-
-void QgsProjectProperties::editDatumTransform()
-{
-  QModelIndexList selectedIndexes = mDatumTransformTableView->selectionModel()->selectedIndexes();
-  if ( selectedIndexes.count() > 0 )
-  {
-    QgsCoordinateReferenceSystem sourceCrs;
-    QgsCoordinateReferenceSystem destinationCrs;
-    int sourceTransform = -1;
-    int destinationTransform = -1;
-    for ( QModelIndexList::const_iterator it = selectedIndexes.constBegin(); it != selectedIndexes.constEnd(); it ++ )
-    {
-      if ( it->column() == QgsDatumTransformTableModel::SourceCrsColumn )
-      {
-        sourceCrs = QgsCoordinateReferenceSystem( mDatumTransformTableModel->data( *it, Qt::DisplayRole ).toString() );
-      }
-      if ( it->column() == QgsDatumTransformTableModel::DestinationCrsColumn )
-      {
-        destinationCrs = QgsCoordinateReferenceSystem( mDatumTransformTableModel->data( *it, Qt::DisplayRole ).toString() );
-      }
-      if ( it->column() == QgsDatumTransformTableModel::SourceTransformColumn )
-      {
-        sourceTransform = mDatumTransformTableModel->data( *it, Qt::UserRole ).toInt();
-      }
-      if ( it->column() == QgsDatumTransformTableModel::DestinationTransformColumn )
-      {
-        destinationTransform = mDatumTransformTableModel->data( *it, Qt::UserRole ).toInt();
-      }
-    }
-    if ( sourceCrs.isValid() && destinationCrs.isValid() &&
-         ( sourceTransform != -1 || destinationTransform != -1 ) )
-    {
-      QgsDatumTransformDialog *dlg = new QgsDatumTransformDialog( sourceCrs, destinationCrs, qMakePair( sourceTransform, destinationTransform ) );
-      if ( dlg->exec() )
-      {
-        QPair< QPair<QgsCoordinateReferenceSystem, int>, QPair<QgsCoordinateReferenceSystem, int > > dt = dlg->selectedDatumTransforms();
-        QgsCoordinateTransformContext context = mDatumTransformTableModel->transformContext();
-        // QMap::insert takes care of replacing existing value
-        context.addSourceDestinationDatumTransform( sourceCrs, destinationCrs, dt.first.second, dt.second.second );
-        mDatumTransformTableModel->setTransformContext( context );
-      }
-    }
-  }
 }
 
 void QgsProjectProperties::cbxWFSPubliedStateChanged( int aIdx )
